lots of changes.
fiddle with lvgl refresh periods and drm to increase frame rate.
most of this is from having drm refresh the display after a vblank,
but some of it is also having input devices run the input timer
directly, and then also trying to refresh the display.

mqtt is wired back in, and wired into luavgl.

the screen blanking status and control is exposed via mqtt. this
could be used with a motion sensor near a display to proactively
turn it on when people are in a space near it.

the lua script can publish telemetry via mqtt when something happens
(eg, a widget is clicked). mqtt cmnds that are not handled by wslv
in c can be passed to a cmnd() function in the lua script to process.
this allows similar (but not the same) control/interaction that
openhasp provides.