if the lua provides a clocktick() function, call it every second.
this isn't great, but it works for now. in the future i'd like to
use something like the kevent ABSTIME handling that was recently
added to kq to fire this off every minute on the minute instead of
around every second like it is now.

i would also like the lua to be able to request the clock tick
instead of wslv assuming there'll be one to call all the time.