avoid segfaults on lua_close by having lv obj gc delete all event handlers.
the reference to the lua state and lobj that lua_lv_event_cb uses
are kept alive with an entry in the lua registry, which in turn
prevents hte gc from running. when lua_close is called, that goes
away so the gc will run, and we should tear all lua related state
on an object down, including lua_lv_event_cb.