try and handle displays where linebytes != the pixels on screen.
the surface go has a resolution of 1800x1200, which is 7200 bytes
per row. the pitch or bytes per line is 7232 or 1808 pixels though.
lvgl drawing a 1800x1200 picture on the framebuffer ends up with a
skewed image because 8 pixels are lost on every line.

cope with this by using the linebytes as the width of the display,
and set physical_hor_res in the display driver settings to the
actual pixels on the display.

this still works on my lenovo yoga s1 (1920x1080 display) where the
pixels and pitch work out the same.