rework screen state handling.
the idle timeout (as specified on the command line) is broken in
half to make two intervals.

when the screen is woken up or touch, the idle timer is reset. when
the first half of the idle time expires the idle state is set to
on, and the timer is restarted. when the second half of the idle
timer expires then the screen is turned off.